Initial commit of `hz`
This *may* hopefully be a tool to search for options to pass to a
binary for which one is seeking "help".

Using fzf, the goal is to search through strings in a binary for
potentional options rather than reading/scanning/skimming all
of the 'help' output, if any, for such options.